+++ CHANGES.rst
@@ -10,6 +10,7 @@ Improvements:
  * Key backup: Implement setup screen (#2198).
  * Key backup: Implement recover screen (#2196).
  * Key backup: Add a dedicated section to settings (#2193).
+ * Key backup: Implement setup reminder (#2211).
  
 Bug fix:
  * Use white scroll bar on dark themes (#2158).